Create WHOISDIGHOST
This is a script built for Mint/Ubuntu... Can confirm it working for the latest versions(as of the date of publication) it should be pretty self-explanatory but I will edit this to have some notes in here when I have a little more time. Honestly, I am massively playing catch up, but just happy to be a part of it.