Update FM.py
Will be making small changes to this just to make it really do what I wanted it too, hope it helps others too. To be honest: had really got more excited about loads of other projects; but as awful as this code is: as it is where I kinda started(outside of web programming anyway); that I owe it to the code to come back and edit it up and fix things once in a while so as to make it not as terrible(and hopefully one day actually good). 

Enjoy.